requirements.txt
google-generativeai
pytubefix
python-dotenv